requirements.txt
pillow==11.1.0
qrcode==7.4.2
